pyproject.toml
[project]
name = "cv-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "openai>=2.9.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
